+++ State_Machine/state_machine.h
@@ -4,9 +4,36 @@
 #include <array>
 #include <iostream>
 #include <string>
-#include "state.h"
+
+template<typename I, typename O>
+class Abstract_State {
+    public: 
+        const std::string name; // name must be unique!
+
+        virtual void entry_behavior(I&, O&) {}
+        virtual void do_behavior(I&, O&) {}
+        virtual void exit_behavior(I&, O&) {}
+        virtual Abstract_State<I, O>& get_next_state(I&) = 0;
+    
+    protected:
+        explicit Abstract_State(std::string name) : name(name){}
 
 
+};
+
+template<typename I, typename O>
+class State_Machine {
+    public: 
+        explicit State_Machine(Abstract_State<I, O>& initial_state, I& input, O& output) : current_state(initial_state), input(input), output(output) {}
+
+    private:
+        Abstract_State<I, O>& current_state;
+        I& input;
+        O& output;
+    
+    public:
+        void iterate();
+};
 
 
 #endif
